+++ core/src/ons/notification.js
@@ -56,7 +56,7 @@ import util from './util';
 const notification = {};
 
 notification._createAlertDialog = function(title, message,
-  buttonLabels, primaryButtonIndex, modifier, animation, callback,
+  buttonLabels, primaryButtonIndex, modifier, animation, id, callback,
   messageIsHTML, cancelable, promptDialog, autofocus, placeholder,
   defaultValue, submitOnEnter, compile) {
 
@@ -72,10 +72,19 @@ notification._createAlertDialog = function(title, message,
     <div class="alert-dialog-footer"></div>
   </ons-alert-dialog>`);
 
+  if (id) {
+    dialogElement.setAttribute('id', id);
+  }
+
   let titleElement = dialogElement.querySelector('.alert-dialog-title');
   let messageElement = dialogElement.querySelector('.alert-dialog-content');
   let footerElement = dialogElement.querySelector('.alert-dialog-footer');
-  let inputElement;
+  let inputElement, result = {};
+
+  result.promise = new Promise((resolve, reject) => {
+    result.resolve = resolve;
+    result.reject = reject;
+  });
 
   modifier = modifier || dialogElement.getAttribute('modifier');
 
@@ -113,6 +122,7 @@ notification._createAlertDialog = function(title, message,
           dialogElement.hide({
             callback: function() {
               callback(inputElement.value);
+              result.resolve(inputElement.value);
               dialogElement.destroy();
               dialogElement = null;
             }
@@ -149,8 +159,10 @@ notification._createAlertDialog = function(title, message,
         callback: function() {
           if (promptDialog) {
             callback(inputElement.value);
+            result.resolve(inputElement.value);
           } else {
             callback(i);
+            result.resolve(i);
           }
           dialogElement.destroy();
           dialogElement = inputElement = buttonElement = null;
@@ -171,8 +183,10 @@ notification._createAlertDialog = function(title, message,
     dialogElement.addEventListener('cancel', function() {
       if (promptDialog) {
         callback(null);
+        result.reject(null);
       } else {
         callback(-1);
+        result.reject(-1);
       }
       setTimeout(function() {
         dialogElement.destroy();
@@ -197,7 +211,7 @@ notification._createAlertDialog = function(title, message,
     dialogElement.setAttribute('modifier', modifier);
   }
 
-  return Promise.resolve(dialogElement);
+  return result.promise;
 };
 
 notification._alertOriginal = function(options) {
@@ -220,6 +234,7 @@ notification._alertOriginal = function(options) {
     0,
     options.modifier,
     options.animation,
+    options.id,
     options.callback,
     !options.message ? true : false,
     false, false, false, '', '', false,
@@ -244,6 +259,9 @@ notification._alertOriginal = function(options) {
  * @param {String} [options.animation]
  *   [en]Animation name. Available animations are "none", "fade" and "slide".[/en]
  *   [ja]アラートダイアログを表示する際のアニメーション名を指定します。"none", "fade", "slide"のいずれかを指定できます。[/ja]
+ * @param {String} [options.id]
+ *   [en]ons-alert-dialog element's ID.[/en]
+ *   [ja]ons-alert-dialog要素のID。[/ja]
  * @param {String} [options.title]
  *   [en]Dialog title. Default is "Alert".[/en]
  *   [ja]アラートダイアログの上部に表示するタイトルを指定します。"Alert"がデフォルトです。[/ja]
@@ -290,6 +308,7 @@ notification._confirmOriginal = function(options) {
     options.primaryButtonIndex,
     options.modifier,
     options.animation,
+    options.id,
     options.callback,
     !options.message ? true : false,
     options.cancelable,
@@ -321,6 +340,9 @@ notification._confirmOriginal = function(options) {
  * @param {String} [options.animation]
  *   [en]Animation name. Available animations are "none", "fade" and "slide".[/en]
  *   [ja]アニメーション名を指定します。"none", "fade", "slide"のいずれかを指定します。[/ja]
+ * @param {String} [options.id]
+ *   [en]ons-alert-dialog element's ID.[/en]
+ *   [ja]ons-alert-dialog要素のID。[/ja]
  * @param {String} [options.title]
  *   [en]Dialog title. Default is "Confirm".[/en]
  *   [ja]ダイアログのタイトルを指定します。"Confirm"がデフォルトです。[/ja]
@@ -376,6 +398,7 @@ notification._promptOriginal = function(options) {
     0,
     options.modifier,
     options.animation,
+    options.id,
     options.callback,
     !options.message ? true : false,
     options.cancelable,
@@ -413,6 +436,9 @@ notification._promptOriginal = function(options) {
  * @param {String} [options.animation]
  *   [en]Animation name. Available animations are "none", "fade" and "slide".[/en]
  *   [ja]アニメーション名を指定します。"none", "fade", "slide"のいずれかを指定します。[/ja]
+ * @param {String} [options.id]
+ *   [en]ons-alert-dialog element's ID.[/en]
+ *   [ja]ons-alert-dialog要素のID。[/ja]
  * @param {String} [options.title]
  *   [en]Dialog title. Default is "Alert".[/en]
  *   [ja]ダイアログのタイトルを指定します。デフォルトは "Alert" です。[/ja]