+++ core/src/ons/notification.spec.js
@@ -7,16 +7,12 @@ describe('ons.notification', () => {
 
   describe('#alert()', () => {
     let dialog,
+      resolvePromise,
       callback = chai.spy();
 
-    beforeEach((done) => {
-      ons.notification.alert({message: 'hoge', modifier: 'fuga', cancelable: true, callback: callback})
-        .then(
-          (dlg) => {
-            dialog = dlg;
-            done();
-          }
-        );
+    beforeEach(() => {
+      resolvePromise = ons.notification.alert({message: 'hoge', modifier: 'fuga', cancelable: true, callback: callback});
+      dialog = document.body.querySelector('ons-alert-dialog');
     });
 
     afterEach(() => {
@@ -30,14 +26,12 @@ describe('ons.notification', () => {
       expect(() => ons.notification.alert()).to.throw(Error);
     });
 
-    it('accepts a \'messageHTML\' parameter', (done) => {
+    it('accepts a \'messageHTML\' parameter', () => {
       let message = '<strong>hoge</strong>';
-      ons.notification.alert({messageHTML: message})
-        .then((dialog) => {
-          expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
-          dialog.destroy();
-          done();
-        });
+      ons.notification.alert({messageHTML: message, id: 'test'});
+      let dialog = document.getElementById('test');
+      expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
+      dialog.destroy();
     });
 
     it('displays an alert dialog', () => {
@@ -56,14 +50,24 @@ describe('ons.notification', () => {
       button.dispatchEvent(event);
       expect(spy).to.have.been.called.once;
     });
+
+    it('resolves to the pressed button index', (done) => {
+      resolvePromise.then(index => {
+        expect(index).to.equal(0);
+        done();
+      });
+
+      dialog.querySelector('button').click();
+    });
   });
 
   describe('#confirm()', () => {
     let dialog,
+      resolvePromise,
       callback = chai.spy();
 
     beforeEach(() => {
-      ons.notification.confirm({message: 'hoge', modifier: 'fuga', cancelable: true, callback: callback});
+      resolvePromise = ons.notification.confirm({message: 'hoge', modifier: 'fuga', cancelable: true, callback: callback});
       dialog = document.body.querySelector('ons-alert-dialog');
     });
 
@@ -78,14 +82,12 @@ describe('ons.notification', () => {
       expect(() => ons.notification.confirm()).to.throw(Error);
     });
 
-    it('accepts a \'messageHTML\' parameter', (done) => {
+    it('accepts a \'messageHTML\' parameter', () => {
       let message = '<strong>hoge</strong>';
-      ons.notification.confirm({messageHTML: message})
-        .then((dialog) => {
-          expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
-          dialog.destroy();
-          done();
-        });
+      ons.notification.confirm({messageHTML: message, id: 'test'});
+      let dialog = document.getElementById('test');
+      expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
+      dialog.destroy();
     });
 
     it('displays an alert dialog', () => {
@@ -110,14 +112,24 @@ describe('ons.notification', () => {
       dialog.dispatchEvent(event);
       expect(callback).to.have.been.called.with(-1);
     });
+
+    it('resolves to the pressed button index', (done) => {
+      resolvePromise.then(index => {
+        expect(index).to.equal(0);
+        done();
+      });
+
+      dialog.querySelector('button').click();
+    });
   });
 
   describe('#prompt()', () => {
     let dialog,
+      resolvePromise,
       callback = chai.spy();
 
     beforeEach(() => {
-      ons.notification.prompt({message: 'hoge', modifier: 'fuga', submitOnEnter: true, cancelable: true, callback: callback});
+      resolvePromise = ons.notification.prompt({message: 'hoge', modifier: 'fuga', submitOnEnter: true, cancelable: true, callback: callback});
       dialog = document.body.querySelector('ons-alert-dialog');
     });
 
@@ -132,14 +144,12 @@ describe('ons.notification', () => {
       expect(() => ons.notification.prompt()).to.throw(Error);
     });
 
-    it('accepts a \'messageHTML\' parameter', (done) => {
+    it('accepts a \'messageHTML\' parameter', () => {
       let message = '<strong>hoge</strong>';
-      ons.notification.prompt({messageHTML: message})
-        .then((dialog) => {
-          expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
-          dialog.destroy();
-          done();
-        });
+      ons.notification.prompt({messageHTML: message, id: 'test'});
+      let dialog = document.getElementById('test');
+      expect(dialog.innerHTML.indexOf(message)).to.be.above(-1);
+      dialog.destroy();
     });
 
     it('displays an alert dialog', () => {
@@ -175,27 +185,35 @@ describe('ons.notification', () => {
       dialog.dispatchEvent(event);
       expect(callback).to.have.been.called.with(null);
     });
+
+    it('resolves to the input value', (done) => {
+      resolvePromise.then(value => {
+        expect(value).to.equal('42');
+        done();
+      });
+
+      dialog.querySelector('input').value = 42;
+      dialog.querySelector('button').click();
+    });
   });
 
   describe('autoStyling', () => {
-    it('adds \'material\' modifier on Android', (done) => {
+    it('adds \'material\' modifier on Android', () => {
       ons.platform.select('android');
-      ons.notification.alert({message: 'test'})
-        .then((dialog) => {
-          expect(dialog.getAttribute('modifier')).to.equal('material');
-          ons.platform.select('');
-          done();
-        });
+      ons.notification.alert({message: 'test', id: 'test'});
+      let dialog = document.getElementById('test');
+      expect(dialog.getAttribute('modifier')).to.equal('material');
+      ons.platform.select('');
+      dialog.remove();
     });
 
     it('removes \'material\' modifier on iOS', () => {
       ons.platform.select('ios');
-      ons.notification.alert({message: 'test'})
-        .then((dialog) => {
-          expect(dialog.getAttribute('modifier')).not.to.equal('material');
-          ons.platform.select('');
-          done();
-        });
+      ons.notification.alert({message: 'test', id: 'test'});
+      let dialog = document.getElementById('test');
+      expect(dialog.getAttribute('modifier')).not.to.equal('material');
+      ons.platform.select('');
+      dialog.remove();
     });
   });
 });